added bookmars

+++ src/v1/modules/bookmark/bookmark.service.ts
@@ -11,14 +11,24 @@ export default class bookMArkService{
 
     public async addBookmark(payload: any){
         try {
+            const check = await this.bookmarkrepo.checkExists(payload)
+
+            if(check){
+                throw new AppError(httpStatus.CONFLICT, "post already bookmarked")
+            }
             await this.bookmarkrepo.save(payload)
         } catch (error: any) {
-            throw new AppError(httpStatus.INTERNAL_SERVER_ERROR, error.message)
+           throw error
         }
     }
 
     public async removeBookmark(id: string){
         try {
+            const bookmark = await this.bookmarkrepo.findById(id);
+            console.log(bookmark)
+            if(!bookmark){
+                throw new AppError(httpStatus.NOT_ACCEPTABLE, "invalid or malformed data passed as bookmark")
+            }
             await this.bookmarkrepo.delete(id)
         } catch (error: any) {
             throw new AppError(httpStatus.INTERNAL_SERVER_ERROR, error.message)
